src/requirements.txt
absl-py==1.4.0
cloudpickle==2.2.1
gym==0.26.2
numpy==1.23.1
nvtx==0.2.5
psutil==5.9.4
pymongo==4.3.3
pynvml==11.4.1
requests==2.28.2
rpyc==5.3.0
scikit_learn==1.2.1
scipy==1.10.0
tabulate==0.9.0
tensorflow==2.11.0
torch==1.12.0
tornado==6.2
tqdm==4.64.1
undecorated==0.3.0
xgboost==1.5.0
scikit-optimize==0.9.0
mxnet==1.9.0
mxnet-cu113==1.9.1
gluonnlp==0.10.0
Pillow==9.4.0